+++ packages/graph/src/gissue-marker.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022 STMicroelectronics and others.
+ * Copyright (c) 2022-2024 STMicroelectronics and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -13,15 +13,8 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-import { DefaultTypes, flatPush, MaybeArray } from '@eclipse-glsp/protocol';
+import { DefaultTypes, flatPush, GIssue, GIssueSeverity, MaybeArray } from '@eclipse-glsp/protocol';
 import { GShapeElement, GShapeElementBuilder } from './gshape-element';
-
-export type GIssueSeverity = 'error' | 'warning' | 'info';
-
-export class GIssue {
-    message: string;
-    severity: GIssueSeverity;
-}
 
 export class GIssueMarker extends GShapeElement {
     static builder(): GIssueMarkerBuilder {